Refactor test file in order to share test with different envs
