+++ test/e2e/tests/local/test_realtime.py
@@ -12,19 +12,11 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import json
-from http import HTTPStatus
-from pathlib import Path
-from typing import List
-
 import cortex as cx
 import pytest
-import yaml
 
-from e2e.expectations import assert_response_expectations, parse_expectations
-from e2e.utils import apis_ready, request_prediction
+import e2e.tests
 
-TEST_APIS_DIR = Path(__file__).parent.parent.parent.parent / "apis"
 TEST_APIS = ["pytorch/iris-classifier", "pytorch/image-classifier-resnet50"]
 DEPLOY_TIMEOUT = 30  # seconds
 
@@ -34,40 +26,6 @@ def client():
     return cx.client("local")
 
 
-def delete_apis(client: cx.Client, api_names: List[str]):
-    for name in api_names:
-        client.delete_api(name)
-
-
 @pytest.mark.parametrize("api", TEST_APIS)
 def test_realtime_apis(client: cx.Client, api: str):
-    api_dir = TEST_APIS_DIR / api
-    with open(str(api_dir / "cortex.yaml")) as f:
-        api_specs = yaml.safe_load(f)
-
-    expectations = None
-    expectations_file = api_dir / "expectations.yaml"
-    if expectations_file.exists():
-        expectations = parse_expectations(str(expectations_file))
-
-    api_names = [api_spec["name"] for api_spec in api_specs]
-    for api_spec in api_specs:
-        client.create_api(api_spec=api_spec, project_dir=api_dir)
-
-    try:
-        assert apis_ready(client=client, api_names=api_names), f"apis {api_names} not ready"
-
-        with open(str(api_dir / "sample.json")) as f:
-            payload = json.load(f)
-
-        api_name = api_names[0]
-        response = request_prediction(client, api_name, payload)
-
-        assert (
-            response.status_code == HTTPStatus.OK
-        ), f"status code: got {response.status_code}, expected {HTTPStatus.OK}"
-
-        if expectations and "response" in expectations:
-            assert_response_expectations(response, expectations["response"])
-    finally:
-        delete_apis(client, api_names)
+    e2e.tests.test_realtime_apis(client=client, api=api, timeout=DEPLOY_TIMEOUT)
